Add a centos kickstart file to create the perfect environment
This kickstart config creates a CentOS environment that is perfect for
fluffy patch building.

It manually creates the OCaml 3.09 environment after the main installation.

This file is available at http://applehq.eu/fluffy-ks.cfg and can be used
when booting a CentOS installation media with the ks=http://... boot option.